# Feature/performance change: lazy parsing