--- a/src/TokenizerInterface.php
+++ b/src/TokenizerInterface.php
@@ -13,8 +13,8 @@ interface TokenizerInterface
     /**
      * Convert the passed message into tokens.
      *
-     * @return Token[]
+     * @return iterable<Token>
      * @throws ParseException
      */
-    public function getTokens(string $message, ControlCharactersInterface $characters): array;
+    public function getTokens(string $message, ControlCharactersInterface $characters): iterable;
 }