We ran into a problem when parsing somewhat sizable EDI files (>25MB) because the parser already broke our 2GB memory limit.

After a bit of fiddling it turned out that the buffering the parser does while in the Tokenizer as well as in `convertTokensToSegments` is quite heavy on the memory. For our 26MB test file the `getTokens` buffer alone added about 1650 MB to the memory, which would not be released until after all segments have been produced. The OOM then broken soon somewhere in `convertTokensToSegments`. 

Since the `convertTokensToSegments` already returns a generator I extended that concept. TokenizerInterface now returns a iterable, and yields tokens on demand. 

`convertTokensToSegments` consumes this generator, pulls tokens on demand, and, instead of first collecting all segments and then yielding them when done, it immediately yields a segment before starting to read the next one. 

This allows the consumer to start processing segments pretty much instantly and moves the whole overhad for the Parser from ~90xFilesize to ~1xFilesize. 

In my tests this improves parsing time for large files by ~20%, since very little time is spend on memory allocation. 

Here are some numbers of benchmarks I ran for this: 
```
# first the current main branch
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:    187.9 MB| peak memory:  2,412.9 MB | after parsing all segments
parsing all segments took 10787 ms

### next with just the first commit, changing Tokenizer
# again 3 runs to check for variance
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.1 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:    833.7 MB | after parsing all segments
parsing all segments took 9173 ms
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:    833.7 MB | after parsing all segments
parsing all segments took 8892 ms
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:    833.7 MB | after parsing all segments
parsing all segments took 9026 ms

### and this is both changes together
# again 3 runs to check for variance
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:     53.7 MB | after parsing all segments
parsing all segments took 8419 ms
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:     53.7 MB | after parsing all segments
parsing all segments took 8211 ms
> php -dmemory_limit=6G bench.php  AN_UNGODLY_AMOUNT_OF_EDI
current memory:     27.0 MB| peak memory:     27.1 MB | after reading file
current memory:     53.7 MB| peak memory:     53.7 MB | after parsing all segments
parsing all segments took 8316 ms
```